Renaming Pioneers with Publishers
Adding the concept of relationships
