+++ constants.php
@@ -1,0 +1,3 @@
+<?php
+
+define('APP_ROOT', __DIR__);
