+++ src/Application/ServiceLocator.php
@@ -5,7 +5,7 @@ namespace CartBooking\Application;
 use CartBooking\Booking\BookingRepository;
 use CartBooking\Lib\Db\Db;
 use CartBooking\Location\LocationRepository;
-use CartBooking\Publisher\PioneerRepository;
+use CartBooking\Publisher\PublisherRepository;
 use CartBooking\Shift\ShiftRepository;
 use Monolog\Logger;
 use Pimple\Container;
@@ -53,7 +53,7 @@ class ServiceLocator
     }
 
     /**
-     * @return PioneerRepository
+     * @return PublisherRepository
      */
     public static function getPioneerRepository()
     {
