+++ src/Publisher/PublisherHydrator.php
@@ -9,11 +9,11 @@
 namespace CartBooking\Publisher;
 
 
-class PioneerHydrator
+class PublisherHydrator
 {
     public function hydrate(array $row)
     {
-        $pioneer = new Pioneer($row['id']);
+        $pioneer = new Publisher($row['id']);
         $pioneer->setFirstName($row['first_name']);
         $pioneer->setLastName($row['last_name']);
         $pioneer->setGender($row['gender']);
@@ -23,6 +23,11 @@ class PioneerHydrator
 //        $pioneer->setStatus($row['status']);
         $pioneer->setPassword((string)$row['password']);
         $pioneer->setInactive($row['inactive'] === 'y' || $row['inactive'] === 1);
+        if (isset($row['relatives'])) {
+            $pioneer->setRelatives(array_map(function (array $row){
+                return $row['publisher_id_2'];
+            }, $row['relatives']));
+        }
         return $pioneer;
     }
 }
